requirements.txt
anthropic>=0.92.0
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
google-auth>=2.29.0
google-api-python-client>=2.126.0
PyYAML>=6.0.1
